# docs/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/users/{id}/password":{"put":{"tags":["user-controller"],"operationId":"changePassword","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/users":{"get":{"tags":["user-controller"],"operationId":"listAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListUser"}}}}}},"post":{"tags":["user-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUser"}}}}}}},"/api/v1/tenants":{"get":{"tags":["tenant-controller"],"operationId":"listActive","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTenant"}}}}}},"post":{"tags":["tenant-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTenantRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTenant"}}}}}}},"/api/v1/tenants/{tenantId}/uploads":{"post":{"tags":["upload-controller"],"operationId":"upload","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringString"}}}}}}},"/api/v1/tenants/{tenantId}/signup-requests":{"get":{"tags":["signup-request-controller"],"operationId":"list","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListSignupRequest"}}}}}},"post":{"tags":["signup-request-controller"],"operationId":"submit","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseSignupRequest"}}}}}}},"/api/v1/tenants/{tenantId}/signup-requests/{id}/reject":{"post":{"tags":["signup-request-controller"],"operationId":"reject","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/tenants/{tenantId}/signup-requests/{id}/approve":{"post":{"tags":["signup-request-controller"],"operationId":"approve","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUser"}}}}}}},"/api/v1/tenants/{tenantId}/products":{"get":{"tags":["product-controller"],"operationId":"list_1","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"activeOnly","in":"query","required":false,"schema":{"type":"boolean","default":true}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListProduct"}}}}}},"post":{"tags":["product-controller"],"operationId":"create_2","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProduct"}}}}}}},"/api/v1/tenants/{tenantId}/orders":{"get":{"tags":["order-controller"],"operationId":"listByTenant_1","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListOrder"}}}}}},"post":{"tags":["order-controller"],"operationId":"place","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseOrder"}}}}}}},"/api/v1/tenants/{tenantId}/charge-requests":{"get":{"tags":["charge-request-controller"],"operationId":"list_2","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListChargeRequest"}}}}}},"post":{"tags":["charge-request-controller"],"operationId":"submit_1","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChargeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseChargeRequest"}}}}}}},"/api/v1/tenants/{tenantId}/charge-requests/{id}/reject":{"post":{"tags":["charge-request-controller"],"operationId":"reject_1","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/tenants/{tenantId}/charge-requests/{id}/approve":{"post":{"tags":["charge-request-controller"],"operationId":"approve_1","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/tenants/{id}/admins":{"get":{"tags":["tenant-controller"],"operationId":"listAdmins","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListUser"}}}}}},"post":{"tags":["tenant-controller"],"operationId":"assignAdmin","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseRefreshResponse"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLoginResponse"}}}}}}},"/api/v1/air/rules":{"get":{"tags":["air-controller"],"operationId":"rules","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListDynamicRule"}}}}}},"post":{"tags":["air-controller"],"operationId":"addRule","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDynamicRule"}}}}}}},"/api/v1/air/incidents/{id}/status":{"post":{"tags":["air-controller"],"operationId":"incidentStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"status","in":"query","required":true,"schema":{"type":"string"}},{"name":"action","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/air/defenses/{key}/enable":{"post":{"tags":["air-controller"],"operationId":"enable","parameters":[{"name":"key","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/air/defenses/{key}/disable":{"post":{"tags":["air-controller"],"operationId":"disable","parameters":[{"name":"key","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/users/{id}":{"get":{"tags":["user-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUser"}}}}}},"delete":{"tags":["user-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}},"patch":{"tags":["user-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUser"}}}}}}},"/api/v1/tenants/{tenantId}/products/{id}":{"get":{"tags":["product-controller"],"operationId":"get_1","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProduct"}}}}}},"delete":{"tags":["product-controller"],"operationId":"delete_1","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}},"patch":{"tags":["product-controller"],"operationId":"update_1","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProduct"}}}}}}},"/api/v1/tenants/{tenantId}/orders/{id}/status":{"patch":{"tags":["order-controller"],"operationId":"updateStatus","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/tenants/{id}":{"get":{"tags":["tenant-controller"],"operationId":"get_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTenant"}}}}}},"delete":{"tags":["tenant-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}},"patch":{"tags":["tenant-controller"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTenantRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTenant"}}}}}}},"/api/v1/users/tenant/{tenantId}":{"get":{"tags":["user-controller"],"operationId":"listByTenant","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListUser"}}}}}}},"/api/v1/users/me":{"get":{"tags":["user-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUser"}}}}}}},"/api/v1/tenants/{tenantId}/uploads/download":{"get":{"tags":["upload-controller"],"operationId":"download","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/v1/tenants/{tenantId}/products/search":{"get":{"tags":["product-controller"],"operationId":"search","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"q","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListProduct"}}}}}}},"/api/v1/tenants/{tenantId}/orders/{id}":{"get":{"tags":["order-controller"],"operationId":"get_2","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseOrder"}}}}}}},"/api/v1/tenants/{tenantId}/orders/my":{"get":{"tags":["order-controller"],"operationId":"listMine","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListOrder"}}}}}}},"/api/v1/tenants/{tenantId}/charge-requests/mine":{"get":{"tags":["charge-request-controller"],"operationId":"mine","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListChargeRequest"}}}}}}},"/api/v1/tenants/managed":{"get":{"tags":["tenant-controller"],"operationId":"listManaged","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTenant"}}}}}}},"/api/v1/tenants/all":{"get":{"tags":["tenant-controller"],"operationId":"listAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTenant"}}}}}}},"/api/v1/notifications/stream":{"get":{"tags":["notification-controller"],"operationId":"stream","responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/SseEmitter"}}}}}}},"/api/v1/notifications/pending-signups":{"get":{"tags":["notification-controller"],"operationId":"pending","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPendingCount"}}}}}}},"/api/v1/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/v1/audit/tenant/{tenantId}":{"get":{"tags":["audit-controller"],"operationId":"byTenant","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListAuditLog"}}}}}}},"/api/v1/audit/resource/{resourceId}":{"get":{"tags":["audit-controller"],"operationId":"byResource","parameters":[{"name":"resourceId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListAuditLog"}}}}}}},"/api/v1/audit/recent":{"get":{"tags":["audit-controller"],"operationId":"recent","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":100}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListAuditLog"}}}}}}},"/api/v1/audit/me":{"get":{"tags":["audit-controller"],"operationId":"mine_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListAuditLog"}}}}}}},"/api/v1/air/incidents":{"get":{"tags":["air-controller"],"operationId":"incidents","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":100}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListMapStringObject"}}}}}}},"/api/v1/air/defenses":{"get":{"tags":["air-controller"],"operationId":"defenses","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringBoolean"}}}}}}},"/api/v1/tenants/{id}/admins/{userId}":{"delete":{"tags":["tenant-controller"],"operationId":"unassignAdmin","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/air/rules/{id}":{"delete":{"tags":["air-controller"],"operationId":"deleteRule","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}}},"components":{"schemas":{"ApiResponseVoid":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object"},"code":{"type":"string"},"message":{"type":"string"}}},"ApiResponseUser":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/User"},"code":{"type":"string"},"message":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"role":{"type":"string","enum":["super_admin","admin","customer"]},"displayName":{"type":"string"},"tenantId":{"type":"string"},"balance":{"type":"integer","format":"int64"},"lastLoginAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"active":{"type":"boolean"},"customer":{"type":"boolean"},"superAdmin":{"type":"boolean"},"admin":{"type":"boolean"}}},"CreateTenantRequest":{"required":["name","slug"],"type":"object","properties":{"name":{"type":"string"},"slug":{"pattern":"^[a-z0-9-]+$","type":"string"},"domain":{"type":"string"},"description":{"type":"string"},"logoUrl":{"type":"string"}}},"ApiResponseTenant":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/Tenant"},"code":{"type":"string"},"message":{"type":"string"}}},"Tenant":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"slug":{"type":"string"},"domain":{"type":"string"},"ownerId":{"type":"string"},"description":{"type":"string"},"logoUrl":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"active":{"type":"boolean"}}},"ApiResponseMapStringString":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","additionalProperties":{"type":"string"}},"code":{"type":"string"},"message":{"type":"string"}}},"CreateSignupRequest":{"required":["password","username"],"type":"object","properties":{"username":{"maxLength":30,"minLength":3,"type":"string"},"password":{"maxLength":72,"minLength":8,"type":"string"},"displayName":{"type":"string"}}},"ApiResponseSignupRequest":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/SignupRequest"},"code":{"type":"string"},"message":{"type":"string"}}},"SignupRequest":{"type":"object","properties":{"id":{"type":"string"},"tenantId":{"type":"string"},"username":{"type":"string"},"displayName":{"type":"string"},"status":{"type":"string"},"rejectReason":{"type":"string"},"reviewedBy":{"type":"string"},"reviewedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"CreateProductRequest":{"required":["name","price","stock"],"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"minimum":0,"type":"integer","format":"int64"},"stock":{"minimum":0,"type":"integer","format":"int32"},"category":{"type":"string"},"imageUrl":{"type":"string"}}},"ApiResponseProduct":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/Product"},"code":{"type":"string"},"message":{"type":"string"}}},"Product":{"type":"object","properties":{"id":{"type":"string"},"tenantId":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"integer","format":"int64"},"stock":{"type":"integer","format":"int32"},"category":{"type":"string"},"imageUrl":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"active":{"type":"boolean"}}},"ItemLine":{"required":["productId","quantity"],"type":"object","properties":{"productId":{"type":"string"},"quantity":{"type":"integer","format":"int32"}}},"PlaceOrderRequest":{"required":["items"],"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/ItemLine"}}}},"ApiResponseOrder":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/Order"},"code":{"type":"string"},"message":{"type":"string"}}},"Order":{"type":"object","properties":{"id":{"type":"string"},"tenantId":{"type":"string"},"customerId":{"type":"string"},"status":{"type":"string","enum":["pending","paid","shipped","cancelled"]},"totalAmount":{"type":"integer","format":"int64"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItem"}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"OrderItem":{"type":"object","properties":{"id":{"type":"string"},"orderId":{"type":"string"},"productId":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"unitPrice":{"type":"integer","format":"int64"}}},"CreateChargeRequest":{"required":["amount"],"type":"object","properties":{"amount":{"maximum":1000000000,"minimum":1,"type":"integer","format":"int64"}}},"ApiResponseChargeRequest":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/ChargeRequest"},"code":{"type":"string"},"message":{"type":"string"}}},"ChargeRequest":{"type":"object","properties":{"id":{"type":"string"},"tenantId":{"type":"string"},"userId":{"type":"string"},"username":{"type":"string"},"amount":{"type":"integer","format":"int64"},"status":{"type":"string"},"rejectReason":{"type":"string"},"reviewedBy":{"type":"string"},"reviewedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseRefreshResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/RefreshResponse"},"code":{"type":"string"},"message":{"type":"string"}}},"RefreshResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"ApiResponseLoginResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/LoginResponse"},"code":{"type":"string"},"message":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"userId":{"type":"string"},"username":{"type":"string"},"role":{"type":"string"},"tenantId":{"type":"string"}}},"ApiResponseDynamicRule":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/DynamicRule"},"code":{"type":"string"},"message":{"type":"string"}}},"DynamicRule":{"type":"object","properties":{"id":{"type":"string"},"ip":{"type":"string"},"method":{"type":"string"},"pathContains":{"type":"string"},"contains":{"type":"string"},"action":{"type":"string"},"source":{"type":"string"},"createdAt":{"type":"integer","format":"int64"}}},"UpdateProductRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"minimum":0,"type":"integer","format":"int64"},"stock":{"minimum":0,"type":"integer","format":"int32"},"category":{"type":"string"},"imageUrl":{"type":"string"},"isActive":{"type":"boolean"}}},"UpdateTenantRequest":{"type":"object","properties":{"name":{"type":"string"},"slug":{"type":"string"},"domain":{"type":"string"},"description":{"type":"string"},"logoUrl":{"type":"string"},"isActive":{"type":"boolean"}}},"ApiResponseListUser":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"code":{"type":"string"},"message":{"type":"string"}}},"ApiResponseListTenant":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Tenant"}},"code":{"type":"string"},"message":{"type":"string"}}},"ApiResponseListSignupRequest":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/SignupRequest"}},"code":{"type":"string"},"message":{"type":"string"}}},"ApiResponseListProduct":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Product"}},"code":{"type":"string"},"message":{"type":"string"}}},"ApiResponseListOrder":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Order"}},"code":{"type":"string"},"message":{"type":"string"}}},"ApiResponseListChargeRequest":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ChargeRequest"}},"code":{"type":"string"},"message":{"type":"string"}}},"SseEmitter":{"type":"object","properties":{"timeout":{"type":"integer","format":"int64"}}},"ApiResponseListPendingCount":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PendingCount"}},"code":{"type":"string"},"message":{"type":"string"}}},"PendingCount":{"type":"object","properties":{"tenantId":{"type":"string"},"tenantName":{"type":"string"},"signupCount":{"type":"integer","format":"int32"},"chargeCount":{"type":"integer","format":"int32"},"orderCount":{"type":"integer","format":"int32"}}},"ApiResponseListAuditLog":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/AuditLog"}},"code":{"type":"string"},"message":{"type":"string"}}},"AuditLog":{"type":"object","properties":{"id":{"type":"string"},"actorId":{"type":"string"},"actorUsername":{"type":"string"},"tenantId":{"type":"string"},"action":{"type":"string"},"resourceType":{"type":"string"},"resourceId":{"type":"string"},"detail":{"type":"string"},"ipAddress":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseListDynamicRule":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/DynamicRule"}},"code":{"type":"string"},"message":{"type":"string"}}},"ApiResponseListMapStringObject":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}},"code":{"type":"string"},"message":{"type":"string"}}},"ApiResponseMapStringBoolean":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","additionalProperties":{"type":"boolean"}},"code":{"type":"string"},"message":{"type":"string"}}}}}}